Control core refactor.
This commit builds up a new in-memory representation of the control
mechanism: a tree of Roles, found in the package workflow.
A role can be of several concrete types to implement different kind of
tree nodes (including template-based iterator nodes), and each role
has either * child Roles or 1 child Task.
The Task is the smallest unit of deployment, and generally there is a
correspondence 1 Task - 1 Executor.
Each Role represents the aggregated state of its child roles and/or
Tasks.
Each Environment does not have a flat list of Tasks any more, but
only a pointer to the root of its Roles tree.
The tree itself is loaded from Configuration by way of an intermediate
YAML representation. This compact YAML representation of the control
tree is unmarshaled and (where needed) template-processed to generate
the full workflow of the environment.