[executor] Support for multiple tasks
The executor now instantiates multiple gRPC clients and multiplexes
incoming mesos messages between them.
Each control message must now include the target descriptor, of which
for the executor the relevant part is the TaskID.
A custom common.TaskCommandInfo is now used instead of mesos.CommandInfo
to describe the task to be launched.
Arguments are not ignored any more if Shell == true, they are instead
appended to the call that's passed to the shell.