[core] Always reuse existing executor if possible
That is, if we get an offer for an agent which already has a running
executor, it is reused. Otherwise, a new executor is launched.
We also use common.TaskCommandInfo instead of mesos.CommandInfo to ship
the task information to the executor, and we set the ControlMode of the
task at mesos-LAUNCH time.