[apricot] A Processor and Repository for Configuration Templates
This looks to create apricot as a standalone binary.
Its primary interface is protos/apricot.proto, and it serves as a shim
daemon (apricot/server.go) for configuration.Service, the non-remote
implementation of the apricot.Service interface.

Future commits:
A remote interface will be apricot/client.go, and this too should
implement apricot.Service.

Most code should be further refactored so that using the interface makes
all calls indistinguishable whether the call is compiled-in, or a gRPC
call implemented in apricot.Client.

coconut should then be cleaned up and Consul KV access code should be
deduplicated. In the end, consul conf should mostly remain as a thin
wrapper around apricot.Client.