[OCTRL-710][core] do not include newline char in JIT config queries
If the config file address was put at the end of JIT command, as below:

o2-qc --config json://path/to/payload

then the regex would include a newline character in the matched string,
which would not pass inputFullRegex when building a query. However,
we would try to retrieve the configuration anyway, because there was no
error check after invoking NewQuery.