[core] remove obsolete workaround to make ERROR state always win
As far as I understand, this was an attempt at fixing the cases when ERROR state is overriden by MIXED, which was fixed in https://github.com/AliceO2Group/Control/pull/535 and https://github.com/AliceO2Group/Control/pull/536 directly in the state arithmetics code.
I don't think this is needed anymore and happy tests seem to confirm it.