[core] log the failed JIT workflow and detector, push DPL out only to IL
As experienced by Uli last night, there was no indication in the message seen in the ECS GUI about which particular workflow failed and which detector does it belong to.
This adds this information and removes DPL command output from that message and sends it to Infologger (saying it can be found there).