OCTRL-899 [core] use SOSOR and EOEOR as run duration for GRP ECS object
Ruben and Chiara need the O2 run duration in GRP ECS objects to be the narrowest timespan which includes CTP start and stop.
Also start of run time should be ready available when the tasks are started.
Because of these two reasons, we should use SOSOR and EOEOR as the timestamps in GRP ECS object.