OCTRL-920 [core] invalidating auto-stop lets the goroutine exit
The goroutine for performing would never exit in case that the auto-stop would be invalidated. Consequently, it would be stuck indefinitely, surpassing the environment's lifetime.
I don't think this was causing any trouble except of just leaking goroutines and obfuscating the state of the application.
We also invalidate the auto-stop in case of going to ERROR, so we do not trigger a spurious transition attempt when it's pointless to do so.