[core] ECS waits until DCS operation completes in case a detector fails
This changes how we consume reply streams of requests to DCS.
In particular, we do not stop consuming events in case we receive a detector error and instead we wait to receive the status of all the other detectors.

As it has been clarified with the DCS team, Ecs2DcsGateway never sends "a pre-closure summary of the current operation" with detector == DCS, despite the documentation in the dcs.proto file.
The latest version of this proto file in the Ecs2DcsGateway repo has the comment removed.

With this commit, we treat an EOF from DCS as the only correct stream termination and do not ever expect a detector == DCS event.

We also remove some reduntant operations, but more refactoring will have to come.

Closes OCTRL-711.