[core] INVARIANT status introduced
INVARIANT status is introduced to fix OCTRL-948. INVARIANT is
overwritten by any other status in status product (X() function). We are
taking account the criticality of the role when merging. If all roles
are noncritical we are returning INVARIANT. If there is any critical
role(s) we are merging and returning their status, which shouldn't be
INVARIANT. All roles except taskRole and callrole are taken as critical