[core] put error in the Error field of events
This swaps Error and Message in two kafka events which deviated from the general convention, where Message contains a general, hard-coded string explaining a moment/step that an environment was at, while Error contains the specific error.
It was agreed with George and this allow COG to look for "cannot create new environment" and "cannot get newly created environment" message to report the global error of an environment creation failure.