[core] split TRG cleanup into two calls for distinguishable kafka events
This commit splits trg.Cleanup() into two calls which should be used subsequently instead of trg.Cleanup.
Thanks to this, the GUIs can listen to kafka events for a specific operationName ("trg.EnsureRunStop") for trigger stop timestamp.
We preferred to avoid having the GUIs hunt for operationStep "perform TRG call: RunStop", which seems to be a message targeted for humans and could be more prone to change, thus breaking behaviour in GUIs.

Fixes OCTRL-1017.