Call ODC STOP only when the partition is in RUNNING
This allows us to stop a run on EPNs when GO_ERROR transition is performed by adding a corresponding hook.
As GO_ERROR can occur with any source state, we make sure to make a STOP call only if the partition is in RUNNING.

At the same time, we will not call ODC STOP when an environment is transitioned due to an ODC partition going into ERROR - a stop does not make sense then.

Additionally, the commit includes minor corrections to a few related logs.

Fixes OCTRL-1036