Rework handling failed GO_ERROR
In this commit we unify the way that a failed GO_ERROR is handled.
We recognize invalid event errors as harmless (ERROR->ERROR is pointless, DONE->ERROR is too late).
Any other case is very much unexpected and we print a visible error and comply with the previous behaviour - setting EROR state manually.

Closes OCTRL-1064.