fix recognizing InvalidEventError in handleFailedGoError
Fixing an issue caused by the fact that I have no idea what I'm doing in Go.
fsm.InvalidEventError implements Error() with a value receiver, which actually allows to provide an error as a value or as a pointer.
However, errors.As() is not smart about it and does not strip the type from references/pointers/etc, so the original version would work only with errors passed as pointers.

This commit modifies the helper to work correctly with errors provided as values, which is the more natural way to use it.
Unfortunately it is not possible to prevent a caller from passing an error pointer.

Let's consider it as a part of OCTRL-1064.