Migrate fill info retrieval to the new LHC plugin
This commit adds a new call in LHC plugin which updates a given environment with the latest known fill information.
Once we validate it, we can proceed with switching off the bkp.RetrieveFillInfo() call.
A necessary change in workflow template will have to follow.

This approach is better, because we don't rely on a man in the middle, thus reduce risks of errors and getting the required info too late (i.e. not before SOR).

Closes OCTRL-1057