be gentler with terminating tasks which fail to reach STANDBY
If a task fails to report STANDBY state upon startup, we now perform the standard TERM, INT, KILL sequence instead of just KILL.
We also rely on Wait() to close relevant pipes and provide us with exit code.
All above is moved into a dedicated method which will be reused in next fixes.